task 2 is corrected
