--- MyWork-test/MyWork-test/src/Azuolas.java
@@ -1,7 +1,0 @@
-public class Azuolas extends Lapuotis {
-    @Override
-    public void turi() {
-        super.turi();
-        System.out.println("Azuolas turi lapus");
-    }
-}
