--- MyWork-test/MyWork-test/src/Berzas.java
@@ -1,7 +1,0 @@
-public class Berzas extends Lapuotis {
-    @Override
-    public void turi() {
-        super.turi();
-        System.out.println("Berzas turi lapus");
-    }
-}
