--- MyWork-test/MyWork-test/src/Egle.java
@@ -1,7 +1,0 @@
-public class Egle extends Spygliuotis {
-    @Override
-    void turi() {
-        super.turi();
-        System.out.println("Egle turi spyglius");
-    }
-}
