--- MyWork-test/MyWork-test/src/Kadagys.java
@@ -1,7 +1,0 @@
-public class Kadagys extends Spygliuotis {
-    @Override
-    void turi() {
-        super.turi();
-        System.out.println("Kadagys turi spyglius");
-    }
-}
