--- MyWork-test/MyWork-test/src/Lapuotis.java
@@ -1,6 +1,0 @@
-public class Lapuotis extends Medis {
-    @Override
-    void turi() {
-        super.turi();
-    }
-}
