--- MyWork-test/MyWork-test/src/Medis.java
@@ -1,5 +1,0 @@
-public class Medis {
-    void turi() {
-        System.out.println("Medis turi kamiena");
-    }
-}
