--- MyWork-test/MyWork-test/src/Pusis.java
@@ -1,7 +1,0 @@
-public class Pusis extends Spygliuotis {
-    @Override
-    void turi() {
-        super.turi();
-        System.out.println("Pusis turi spyglius");
-    }
-}
