--- MyWork-test/MyWork-test/src/Spygliuotis.java
@@ -1,6 +1,0 @@
-public class Spygliuotis extends Medis {
-    @Override
-    void turi() {
-        super.turi();
-    }
-}
