+++ MyWork-test/MyWork-test/src/Task1.java
@@ -1,4 +1,4 @@
-public class Main {
+public class Task1 {
     public static void main(String[] args) {
         Integer[] numbersArray = {1, 2, 3};
         String[] stringsArray = {"Hello", "world"};
@@ -6,8 +6,6 @@ public class Main {
         printArray(stringsArray);
         System.out.println();
 
-        Azuolas azuolas = new Azuolas();
-        azuolas.turi();
     }
 
     private static <T> void printArray(T[] array) {
