+++ MyWork-test/MyWork-test/src/task2_medziai/Medis.java
@@ -1,0 +1,5 @@
+package task2_medziai;
+
+public abstract class Medis {
+    public abstract String turi();
+}
