+++ MyWork-test/MyWork-test/src/task2_medziai/rusys/Berzas.java
@@ -1,0 +1,10 @@
+package task2_medziai.rusys;
+
+import task2_medziai.tipai.Lapuotis;
+
+public class Berzas extends Lapuotis {
+    @Override
+    public String turi() {
+        return "Berzas turi lapus";
+    }
+}
