+++ MyWork-test/MyWork-test/src/task2_medziai/rusys/Egle.java
@@ -1,0 +1,10 @@
+package task2_medziai.rusys;
+
+import task2_medziai.tipai.Spygliuotis;
+
+public class Egle extends Spygliuotis {
+    @Override
+    public String turi() {
+        return "Egle turi spyglius";
+    }
+}
