+++ MyWork-test/MyWork-test/src/task2_medziai/rusys/Kadagys.java
@@ -1,0 +1,10 @@
+package task2_medziai.rusys;
+
+import task2_medziai.tipai.Spygliuotis;
+
+public class Kadagys extends Spygliuotis {
+    @Override
+    public String turi() {
+        return "Kadagys turi spyglius";
+    }
+}
