+++ MyWork-test/MyWork-test/src/task2_medziai/rusys/Pusis.java
@@ -1,0 +1,10 @@
+package task2_medziai.rusys;
+
+import task2_medziai.tipai.Spygliuotis;
+
+public class Pusis extends Spygliuotis {
+    @Override
+    public String turi() {
+        return "Pusis turi spyglius";
+    }
+}
