+++ MyWork-test/MyWork-test/src/task2_medziai/tipai/Spygliuotis.java
@@ -1,0 +1,6 @@
+package task2_medziai.tipai;
+
+import task2_medziai.Medis;
+
+public abstract class Spygliuotis extends Medis {
+}
